Move backlight into its own package
