+++ doc/api-documentation.rst
@@ -5,7 +5,15 @@ API Documentation
     :undoc-members:
     :show-inheritance:
 
-.. inheritance-diagram:: luma.core.device luma.core.mixin luma.core.virtual luma.lcd.device
+.. inheritance-diagram:: luma.core.device luma.core.mixin luma.core.virtual luma.lcd.device luma.lcd.aux
+
+:mod:`luma.lcd.aux`
+"""""""""""""""""""
+.. automodule:: luma.lcd.aux
+    :members:
+    :inherited-members:
+    :undoc-members:
+    :show-inheritance:
 
 :mod:`luma.lcd.device`
 """"""""""""""""""""""
