+++ luma/lcd/device.py
@@ -210,31 +210,3 @@ class st7735(device):
         self._serial_interface.command(cmd)
         if len(args) > 0:
             self._serial_interface.data(list(args))
-
-
-class backlight(object):
-    """
-    Controls a backlight (assumed to be on GPIO 18 (PWMCLK0))
-    """
-    def __init__(self, gpio=None, bcm_LIGHT=18):
-        self._bcm_LIGHT = bcm_LIGHT
-        self._gpio = gpio or self.__rpi_gpio__()
-        self._gpio.setmode(self._gpio.BCM)
-        self._gpio.setup(self._bcm_LIGHT, self._gpio.OUT)
-        self.enable(True)
-
-    def enable(self, value):
-        """
-        Switches on the backlight when ``True`` supplied, else ``False``
-        switches it off
-        """
-        assert(value in [True, False])
-        self._gpio.output(self._bcm_LIGHT,
-                          self._gpio.LOW if value else self._gpio.HIGH)
-
-    def __rpi_gpio__(self):
-        # RPi.GPIO _really_ doesn't like being run on anything other than
-        # a Raspberry Pi... this is imported here so we can swap out the
-        # implementation for a mock
-        import RPi.GPIO
-        return RPi.GPIO
